test: move expire-flag and truncate_text tests into test_utils.py with fixed_client marker
Also cover truncate_text's default max_length and the single-long-token path.
