+++ tests/test_utils.py
@@ -1,6 +1,7 @@
-from datetime import datetime
+from datetime import datetime, timedelta, timezone
 import warnings
 import pytest
+from redis.exceptions import DataError
 from redis.utils import (
     DEFAULT_RESP_VERSION,
     check_protocol_version,
@@ -9,6 +10,8 @@ from redis.utils import (
     deprecated_args,
     experimental_method,
     experimental_args,
+    extract_expire_flags,
+    truncate_text,
 )
 
 
@@ -184,3 +187,68 @@ class TestExperimentalArgs:
             result = func_no_args()
             assert result == "no_args"
             assert len(w) == 0
+
+
+@pytest.mark.fixed_client
+class TestExtractExpireFlags:
+    def test_no_flags_returns_empty(self):
+        assert extract_expire_flags() == []
+
+    def test_ex_as_int(self):
+        assert extract_expire_flags(ex=10) == ["EX", 10]
+
+    def test_ex_as_timedelta_uses_total_seconds(self):
+        assert extract_expire_flags(ex=timedelta(minutes=1)) == ["EX", 60]
+
+    def test_ex_as_digit_string(self):
+        assert extract_expire_flags(ex="60") == ["EX", 60]
+
+    def test_ex_invalid_raises(self):
+        with pytest.raises(DataError):
+            extract_expire_flags(ex="not-a-number")
+
+    def test_px_as_int(self):
+        assert extract_expire_flags(px=500) == ["PX", 500]
+
+    def test_px_as_timedelta_uses_milliseconds(self):
+        assert extract_expire_flags(px=timedelta(seconds=2)) == ["PX", 2000]
+
+    def test_exat_as_int(self):
+        assert extract_expire_flags(exat=1700000000) == ["EXAT", 1700000000]
+
+    def test_exat_as_datetime_uses_timestamp(self):
+        when = datetime(2023, 11, 14, 22, 13, 20, tzinfo=timezone.utc)
+        assert extract_expire_flags(exat=when) == ["EXAT", int(when.timestamp())]
+
+    def test_pxat_as_int(self):
+        assert extract_expire_flags(pxat=1700000000000) == ["PXAT", 1700000000000]
+
+    def test_ex_takes_precedence_over_px(self):
+        # The flags are checked in order; ex wins when several are given.
+        assert extract_expire_flags(ex=5, px=999) == ["EX", 5]
+
+
+@pytest.mark.fixed_client
+class TestTruncateText:
+    def test_short_text_unchanged(self):
+        assert truncate_text("hi") == "hi"
+
+    def test_long_text_gets_ellipsis(self):
+        result = truncate_text("hello world this is long", 10)
+        assert result == "hello..."
+        assert len(result) <= 10
+
+    def test_default_max_length_truncates_long_text(self):
+        # Every real call site relies on the default width of 100; make sure a
+        # text longer than that is shortened without an explicit width argument.
+        text = "word " * 40  # 200 characters
+        result = truncate_text(text)
+        assert len(result) <= 100
+        assert result.endswith("...")
+
+    def test_single_long_token_collapses_to_placeholder(self):
+        # A single unbroken token wider than ``max_length``: ``textwrap.shorten``
+        # has no whitespace to truncate on and does not split within a word (its
+        # ``break_long_words`` does not apply to this case), so it drops the word
+        # entirely and returns just the placeholder.
+        assert truncate_text("a" * 200) == "..."
